fix(ci): Run pyright postinstall scripts manually
Pyright postinstall scripts are as follow (stripped)

```json
    "scripts": {
        "postinstall": "node ./build/skipBootstrap.js || npm run install:others",
        "install:others": "cross-env SKIP_LERNA_BOOTSTRAP=yes lerna exec --no-bail npm install",
    }
```

When running `pybun install` on the pyright repository, bun execute the postinstall
script and replace `npm run install:others` with `bun install:others`.

But since we use pybun, we don't have a `bun` CLI available so the `bun install:others`
command fail.

To fix this, I rewrite the postinstall script to make it use `pybun`
instead of `bun`.